Refactor: remove class to utils/data.ts
